Use size_t rather than unsigned int in previous patch.

+++ src/lalr.h
@@ -56,7 +56,7 @@ void lalr_free (void);
    together and GOTO_MAP[I - NTOKENS] is the index in FROM_STATE and
    TO_STATE of the first of them.  */
 
-typedef unsigned int goto_number;
+typedef size_t goto_number;
 
 extern goto_number *goto_map;
 extern state_number *from_state;
